+++ src/main/kotlin/com/pestphp/pest/configuration/PestRunConfigurationHandler.kt
@@ -2,12 +2,13 @@ package com.pestphp.pest.configuration
 
 import com.intellij.execution.ExecutionException
 import com.intellij.openapi.project.Project
-import com.jetbrains.php.composer.lib.ComposerLibraryManager
 import com.jetbrains.php.config.commandLine.PhpCommandSettings
 import com.jetbrains.php.testFramework.run.PhpTestRunConfigurationHandler
 import com.pestphp.pest.toPestTestRegex
 
 class PestRunConfigurationHandler : PhpTestRunConfigurationHandler {
+    var rootPath: String? = null
+
     companion object {
         @JvmField val instance = PestRunConfigurationHandler()
     }
@@ -72,9 +73,7 @@ class PestRunConfigurationHandler : PhpTestRunConfigurationHandler {
         }
         val pathMapper = phpCommandSettings.pathProcessor.createPathMapper(project)
 
-        val rootPath =
-            ComposerLibraryManager.getInstance(project).findVendorDirForUpsource()?.parent?.path
-                ?: workingDirectory
+        val rootPath = this.rootPath ?: workingDirectory
 
         phpCommandSettings.addPathArgument(file)
         phpCommandSettings.addArgument(